requirements.txt
qrcode
Pillow
python-dotenv
pandas
streamlit
opencv-python
pyzbar
numpy
